APPEALS-44441. Added Custom and Scenario Seed components and jest tests
