+++ client/COPY.json
@@ -1481,5 +1481,6 @@
   "CASE_DISTRIBUTION_LEVER_HISTORY_UPDATED_VALUE": "Updated Value",
   "CASE_DISTRIBUTION_STATIC_LEVERS_VALUES": "Values",
   "TEST_SEEDS_ALERT_MESSAGE": " in progress",
-  "TEST_SEEDS_RUN_SEEDS": "Scenario Seeds"
+  "TEST_SEEDS_RUN_SEEDS": "Scenario Seeds",
+  "TEST_SEEDS_CUSTOM_SEEDS": "Custom Seeds"
 }
