+++ client/app/index.js
@@ -59,6 +59,7 @@ import MPISearch from 'app/mpi/MPISearch';
 import Admin from 'app/admin';
 import CaseDistribution from 'app/caseDistribution';
 import CaseDistributionTest from 'app/caseDistribution/test';
+import TestSeeds from 'app/testSeeds';
 import uuid from 'uuid';
 
 const COMPONENTS = {
@@ -95,7 +96,8 @@ const COMPONENTS = {
   MPISearch,
   Admin,
   CaseDistribution,
-  CaseDistributionTest
+  CaseDistributionTest,
+  TestSeeds
 };
 
 const componentWrapper = (component) => (props, railsContext, domNodeId) => {
@@ -180,7 +182,8 @@ const componentWrapper = (component) => (props, railsContext, domNodeId) => {
         './explain/index',
         './mpi/MPISearch',
         './admin/index',
-        './caseDistribution/index'
+        './caseDistribution/index',
+        './testSeeds/index'
       ],
       () => renderApp(component)
     );
