+++ client/app/testSeeds/pages/TestSeedsApp.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import TestSeedsWrapper from '../components/TestSeedsWrapper';
+
+class TestSeedsApp extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <div>
+        <div>
+          <TestSeedsWrapper />
+        </div>
+      </div>
+    );
+
+  }
+}
+
+export default TestSeedsApp;
+
