+++ client/constants/CUSTOM_SEEDS.json
@@ -1,0 +1,6 @@
+{
+  "seed-ama-aod-hearing": "Seed AMA AOD Hearing",
+  "seed-ama-non-aod-hearing": "Seed AMA NON AOD Hearing",
+  "seed-legacy-cases": "Seed Legacy Cases",
+  "seed-ama-direct-review": "Seed AMA Direct Review"
+}
